+++ web-wallet/src/lib/navigation/logout.js
@@ -1,0 +1,20 @@
+import { goto } from "$app/navigation";
+
+import { walletStore } from "$lib/stores";
+
+/**
+ * Logs out the user.
+ * If `isForced` is set to `true`, a querystring
+ * parameter is added to instruct the home page
+ * to add an explanation about the forced logout.
+ *
+ * @param {boolean} isForced
+ * @returns {ReturnType<goto>}
+ */
+const logout = isForced => {
+	walletStore.reset();
+
+	return goto(`/${isForced ? "forced-logout" : ""}`);
+};
+
+export default logout;